# api-gateway/src/main/resources/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: E-Commerce Platform API
  description: |
    E-Commerce Microservices Platform API documentation.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Services
    - **Customer Service**: User registration, authentication, membership
    - **Product Service**: Product catalog, search, categories
    - **Order Service**: Shopping cart, order management
    - **Payment Service**: Payment processing
    - **Logistics Service**: Shipment tracking
    - **Sales Service**: Promotions and coupons
    - **Admin Portal**: Back-office management
  version: 1.0.0
  contact:
    name: E-Commerce Team
    email: support@ecommerce.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.ecommerce.com
    description: Production
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
...
  /api/admin/auth/login:
    post:
      tags: [Admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAuthResponse'
  /api/admin/products:
    get:
      tags: [Admin]
      summary: List products (admin)
      security:
        - adminAuth: []
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminProductSummary'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthResponse:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        email:
          type: string
        token:
          type: string
        expiresIn:
          type: integer
...
    AdminLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AdminAuthResponse:
      type: object
      properties:
        adminId:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        token:
          type: string
    AdminProductSummary:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        price:
          type: number
        stock:
          type: integer
        active:
          type: boolean